Create allZerosBest.cpp
Submission for readable solution, problem with previous attempts were i was only considering the identical case where the two items were adjacent instead of the whole list. Complexity is O(n^2) as this is converting the sequence to zeros whereas the solution that gives number of operations can be done in O(1).